LayoutTests/svg/dynamic-updates/SVGFEGaussianBlurElement-dom-stdDeviation-call-expected.txt

CONSOLE MESSAGE: line 40: TypeError: 'undefined' is not a function (evaluating 'blurElement.setStdDeviation(10, 10)')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stdDeviation' property of the SVGFEGaussianBlurElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL blurElement.stdDeviationX.baseVal should be 5. Threw exception TypeError: 'undefined' is not an object (evaluating 'blurElement.stdDeviationX.baseVal')
FAIL blurElement.stdDeviationY.baseVal should be 5. Threw exception TypeError: 'undefined' is not an object (evaluating 'blurElement.stdDeviationY.baseVal')
